add check if wine id exist otherwise error message

+++ Server/routes/deleteWine.py
@@ -9,6 +9,20 @@ def delete_wine():
 
     con = get_db_connection()
     cursor = con.cursor(dictionary=True)
+
+    # checking if wine exist
+    check_sql = "SELECT * FROM wines WHERE wine_id = %s"
+    check_value= (wine_id,)
+    
+    cursor.execute(check_sql, check_value)
+
+    wine_exists = cursor.fetchone()
+    if not wine_exists:
+        cursor.close()
+        con.close()
+        return jsonify({
+            "error": "wine id does not exist"
+        }), 404
    
     # insert admin into admins table
     sql = "DELETE FROM wines WHERE wine_id = %s"
